# Website/storage/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: HDC API Services
  version: 1.0.0
  description: API for managing HDC
paths:
...
  /api/v1/assets/{id}:
    get:
      summary: Get a specific asset
      operationId: assets.show
      tags:
        - Assets
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the asset
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '404':
          description: Asset not found
    put:
      summary: Update a specific asset
      operationId: assets.update
      tags:
        - Assets
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the asset
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Asset'
      responses:
        '200':
          description: Asset updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '404':
          description: Asset not found
    delete:
      summary: Delete a specific asset
      operationId: assets.destroy
      tags:
        - Assets
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the asset
          schema:
            type: integer
      responses:
        '204':
          description: Asset deleted successfully
        '404':
          description: Asset not found
...
  /api/v1/clients/{id}:
    get:
      summary: Get a specific client
      operationId: clients.show
      tags:
        - Clients
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the client
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: Client not found
    put:
      summary: Update a specific client
      operationId: clients.update
      tags:
        - Clients
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the client
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '200':
          description: Client updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: Client not found
    delete:
      summary: Delete a specific client
      operationId: clients.destroy
      tags:
        - Clients
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the client
          schema:
            type: integer
      responses:
        '204':
          description: Client deleted successfully
        '404':
          description: Client not found
...
  /api/v1/permissions/{id}:
    get:
      summary: Get a specific permission
      operationId: permissions.show
      tags:
        - Permissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the permission
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '404':
          description: Permission not found
    put:
      summary: Update a specific permission
      operationId: permissions.update
      tags:
        - Permissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the permission
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
      responses:
        '200':
          description: Permission updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '404':
          description: Permission not found
    delete:
      summary: Delete a specific permission
      operationId: permissions.destroy
      tags:
        - Permissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the permission
          schema:
            type: integer
      responses:
        '204':
          description: Permission deleted successfully
        '404':
          description: Permission not found
...
  /api/v1/user-permissions/{id}:
    get:
      summary: Get a specific user permission
      operationId: userPermissions.show
      tags:
        - UserPermissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user permission
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPermission'
        '404':
          description: User permission not found
    put:
      summary: Update a specific user permission
      operationId: userPermissions.update
      tags:
        - UserPermissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user permission
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPermission'
      responses:
        '200':
          description: User permission updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPermission'
        '404':
          description: User permission not found
    delete:
      summary: Delete a specific user permission
      operationId: userPermissions.destroy
      tags:
        - UserPermissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user permission
          schema:
            type: integer
      responses:
        '204':
          description: User permission deleted successfully
        '404':
          description: User permission not found
...
  /api/v1/vendors/{id}:
    get:
      summary: Get a specific vendor
      operationId: vendors.show
      tags:
        - Vendors
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the vendor
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '404':
          description: Vendor not found
    put:
      summary: Update a specific vendor
      operationId: vendors.update
      tags:
        - Vendors
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the vendor
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vendor'
      responses:
        '200':
          description: Vendor updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '404':
          description: Vendor not found
    delete:
      summary: Delete a specific vendor
      operationId: vendors.destroy
      tags:
        - Vendors
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the vendor
          schema:
            type: integer
      responses:
        '204':
          description: Vendor deleted successfully
        '404':
          description: Vendor not found
...
  /api/v1/types/{id}:
    get:
      summary: Get a specific type
      operationId: types.show
      tags:
        - Types
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the type
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Type'
        '404':
          description: Type not found
    put:
      summary: Update a specific type
      operationId: types.update
      tags:
        - Types
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the type
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Type'
      responses:
        '200':
          description: Type updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Type'
        '404':
          description: Type not found
    delete:
      summary: Delete a specific type
      operationId: types.destroy
      tags:
        - Types
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the type
          schema:
            type: integer
      responses:
        '204':
          description: Type deleted successfully
        '404':
          description: Type not found
...
  /api/v1/roles/{id}:
    get:
      summary: Get a specific role
      operationId: roles.show
      tags:
        - Roles
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the role
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: Role not found
    put:
      summary: Update a specific role
      operationId: roles.update
      tags:
        - Roles
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the role
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: Role not found
    delete:
      summary: Delete a specific role
      operationId: roles.destroy
      tags:
        - Roles
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the role
          schema:
            type: integer
      responses:
        '204':
          description: Role deleted successfully
        '404':
          description: Role not found
...
  /api/v1/status/{id}:
    get:
      summary: Get a specific status
      operationId: status.show
      tags:
        - Status
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the status
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '404':
          description: Status not found
    put:
      summary: Update a specific status
      operationId: status.update
      tags:
        - Status
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the status
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Status'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '404':
          description: Status not found
    delete:
      summary: Delete a specific status
      operationId: status.destroy
      tags:
        - Status
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the status
          schema:
            type: integer
      responses:
        '204':
          description: Status deleted successfully
        '404':
          description: Status not found
...
  /api/v1/staff/staffs/{id}:
    get:
      summary: Get a specific staff member
      operationId: staff.show
      tags:
        - Staff
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff member
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '404':
          description: Staff member not found
    put:
      summary: Update a specific staff member
      operationId: staff.update
      tags:
        - Staff
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff member
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Staff'
      responses:
        '200':
          description: Staff member updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '404':
          description: Staff member not found
    delete:
      summary: Delete a specific staff member
      operationId: staff.destroy
      tags:
        - Staff
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff member
          schema:
            type: integer
      responses:
        '204':
          description: Staff member deleted successfully
        '404':
          description: Staff member not found
...
  /api/v1/staff/details/{id}:
    get:
      summary: Get a specific staff detail
      operationId: staffDetails.show
      tags:
        - StaffDetails
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff detail
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffDetail'
        '404':
          description: Staff detail not found
    put:
      summary: Update a specific staff detail
      operationId: staffDetails.update
      tags:
        - StaffDetails
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff detail
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffDetail'
      responses:
        '200':
          description: Staff detail updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffDetail'
        '404':
          description: Staff detail not found
    delete:
      summary: Delete a specific staff detail
      operationId: staffDetails.destroy
      tags:
        - StaffDetails
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff detail
          schema:
            type: integer
      responses:
        '204':
          description: Staff detail deleted successfully
        '404':
          description: Staff detail not found
...
  /api/v1/staff/salaries/{id}:
    get:
      summary: Get a specific staff salary
      operationId: staffSalaries.show
      tags:
        - StaffSalaries
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff salary
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffSalary'
        '404':
          description: Staff salary not found
    put:
      summary: Update a specific staff salary
      operationId: staffSalaries.update
      tags:
        - StaffSalaries
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff salary
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffSalary'
      responses:
        '200':
          description: Staff salary updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffSalary'
        '404':
          description: Staff salary not found
    delete:
      summary: Delete a specific staff salary
      operationId: staffSalaries.destroy
      tags:
        - StaffSalaries
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the staff salary
          schema:
            type: integer
      responses:
        '204':
          description: Staff salary deleted successfully
        '404':
          description: Staff salary not found
...
  /api/v1/projects/{id}:
    get:
      summary: Get a specific project
      operationId: projects.show
      tags:
        - Projects
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    put:
      summary: Update a specific project
      operationId: projects.update
      tags:
        - Projects
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    delete:
      summary: Delete a specific project
      operationId: projects.destroy
      tags:
        - Projects
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project
          schema:
            type: integer
      responses:
        '204':
          description: Project deleted successfully
        '404':
          description: Project not found
...
  /api/v1/projects/details/{id}:
    get:
      summary: Get a specific project detail
      operationId: projectDetails.show
      tags:
        - ProjectDetails
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project detail
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '404':
          description: Project detail not found
    put:
      summary: Update a specific project detail
      operationId: projectDetails.update
      tags:
        - ProjectDetails
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project detail
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectDetail'
      responses:
        '200':
          description: Project detail updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '404':
          description: Project detail not found
    delete:
      summary: Delete a specific project detail
      operationId: projectDetails.destroy
      tags:
        - ProjectDetails
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project detail
          schema:
            type: integer
      responses:
        '204':
          description: Project detail deleted successfully
        '404':
          description: Project detail not found
...
  /api/v1/project/project-permissions/{id}:
    get:
      summary: Get a specific project permission
      operationId: projectPermissions.show
      tags:
        - ProjectPermissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project permission
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPermission'
        '404':
          description: Project permission not found
    put:
      summary: Update a specific project permission
      operationId: projectPermissions.update
      tags:
        - ProjectPermissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project permission
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectPermission'
      responses:
        '200':
          description: Project permission updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPermission'
        '404':
          description: Project permission not found
    delete:
      summary: Delete a specific project permission
      operationId: projectPermissions.destroy
      tags:
        - ProjectPermissions
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the project permission
          schema:
            type: integer
      responses:
        '204':
          description: Project permission deleted successfully
        '404':
          description: Project permission not found
components:
  schemas:
    Asset:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        name:
          type: string
        price:
          type: number
        buy_date:
          type: string
          format: date
        condition:
          type: string
        grant_staff:
          type: string
        grant_date:
          type: string
          format: date
        status:
          type: string
      required:
        - type
        - name
        - price
        - condition
        - status
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          format: phone
      required:
        - name
    Permission:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
      required:
        - name
    UserPermission:
      type: object
      properties:
        id:
          type: integer
        staff_id:
          type: string
        permission_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - staff_id
        - permission_id
    Vendor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: integer
        address:
          type: string
        type_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
        - phone
        - address
        - type_id
    Type:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
    Status:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
    Staff:
      type: object
      properties:
        id:
          type: integer
        staff_id:
          type: integer
        fullname:
          type: string
        phone:
          type: integer
          format: phone
        email:
          type: string
          format: email
        role_id:
          type: integer
        password:
          type: string
    StaffDetail:
      type: object
      properties:
        id:
          type: integer
        staff_id:
          type: integer
        bank_number:
          type: integer
        bank_name:
          type: string
        social_insurance:
          type: string
        join_date:
          type: string
          format: date
      required:
        - staff_id
    StaffSalary:
      type: object
      properties:
        id:
          type: integer
        staff_id:
          type: integer
        amount:
          type: number
        date:
          type: string
          format: date
      required:
        - staff_id
        - amount
        - date
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cost:
          type: integer
        status:
          type: string
        client_id:
          type: integer
      required:
        - name
        - cost
        - status
        - client_id
    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        vendor_id:
          type: integer
        name:
          type: string
        type:
          type: string
        cost:
          type: number
        internal_cost:
          type: number
        real_cost:
          type: number
        status:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - project_id
        - vendor_id
        - name
        - type
        - cost
        - internal_cost
        - real_cost
        - status
    ProjectPermission:
      type: object
      properties:
        id:
          type: integer
        staff_id:
          type: string
        project_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - staff_id
        - project_id